+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [master](https://github.com/avbentem/airtime-calculator/compare/v1.2.1...master) (future)
+
+### Fixes
+
+- Improve documentation.
+
 ## [1.2.1](https://github.com/avbentem/airtime-calculator/compare/v1.2.0...v1.2.1) (2020-08-08)
 
 ### New features