+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Lower minimum LoRaWAN overhead size to 12 bytes, as FPort is only given when an application
   payload is present.
+- For CN470-510, add maximum dwell time and note about Listen Before Talk, as per [upcoming
+  regulations](https://lora-alliance.org/sites/default/files/2020-06/rp_2-1.0.1.pdf#page=51).
 - Improve documentation.
 
 ## [1.2.1](https://github.com/avbentem/airtime-calculator/compare/v1.2.0...v1.2.1) (2020-08-08)