+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [master](https://github.com/avbentem/airtime-calculator/compare/v1.2.1...master) (future)
+## [1.3.0](https://github.com/avbentem/airtime-calculator/compare/v1.2.1...v1.3.0) (2020-09-05)
 
 ### New features